refactored the entire code so that the cnp is of type long long

+++ RDS-code/contract.h
@@ -14,12 +14,12 @@ void createcontract();
 void select_service(contract *con);
 void select_type(contract *con);
 void select_years(contract* con);
-void read_cnp(long* cnp);
+void read_cnp(long long* cnp);
 char* read_name();
 char* read_first_name();
 char* read_address();
-int confirm_data(contract con, long cnp, char* name, char* first_name, char* address);
-void submit_data(contract option, long cnp, char* name, char* first_name, char* address);
+int confirm_data(contract con, long long cnp, char* name, char* first_name, char* address);
+void submit_data(contract option, long long cnp, char* name, char* first_name, char* address);
 
 #endif
